requirements.txt
# Framework
fastapi==0.104.0
uvicorn[standard]==0.23.2

# API Request/Response Handling
pydantic==2.4.2
pydantic_settings==2.0.3
python-multipart==0.0.6

# HTTP Clients
httpx==0.25.0
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# Groq API Integration
groq==0.4.1

# Google APIs
google-api-python-client==2.100.0
google-auth==2.22.0
google-auth-oauthlib==1.1.0


# OpenAPI and Documentation
starlette==0.27.0

